# services/customer-gw-service/docs/swagger.yaml
basePath: /
definitions:
  dto.AuthTokenResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at_unix_ms:
        type: integer
      refresh_token:
        type: string
      refresh_token_expires_at_unix_ms:
        type: integer
    type: object
  dto.DeactivateResponse:
    properties:
      success:
        type: boolean
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      auth_token:
        $ref: '#/definitions/dto.AuthTokenResponse'
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.LogoutResponse:
    properties:
      success:
        type: boolean
    type: object
  dto.ProfileResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  dto.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.UpdateProfileRequest:
    properties:
      full_name:
        type: string
      phone_number:
        type: string
    type: object
  dto.UpdateProfileResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.UserResponse:
    properties:
      created_at_unix_ms:
        type: integer
      deleted_at_unix_ms:
        type: integer
      disabled:
        type: boolean
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone_number:
        type: string
      updated_at_unix_ms:
        type: integer
    type: object
  response.Meta:
    properties:
      page:
        description: Current page number
        type: integer
      per_page:
        description: Items per page
        type: integer
      total:
        description: Total number of items
        type: integer
      total_pages:
        description: Total number of pages
        type: integer
    type: object
  response.Response:
    properties:
      data:
        description: Response payload
      errors:
        description: Error details
      message:
        description: Human-readable message
        type: string
      meta:
        allOf:
        - $ref: '#/definitions/response.Meta'
        description: Metadata (pagination, etc.)
      status:
        description: HTTP status code
        type: integer
      success:
        description: Indicates if the request was successful
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: support@stockanalyzr.com
    name: API Support
  description: HTTP API Gateway for Stockanalyzr - manages customer operations with
    JWT authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Customer Gateway Service API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access/refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the current access token
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LogoutResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, and profile details
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - auth
  /api/v1/users/me:
    delete:
      consumes:
      - application/json
      description: Deactivate the authenticated user's account
      produces:
      - application/json
      responses:
        "200":
          description: Account deactivated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DeactivateResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Deactivate user account
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile information
      parameters:
      - description: Profile update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateProfileResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"